Introduce `allow_shutoff` and `minimum_duty_cycle`
This allows for control of lower fan values. They are enforced before
the desired duty cycle is written to the handle.

If the fan allows to be shut off, any value below `minimum_duty_cycle`
will make it to turn completely off, that is set the handle to 0.

If the fan does not allow to be shut off, any value below
`minimum_duty_cycle` is set to that minimum value before writing to the
handle.